requirements.txt
nemo_toolkit[asr]>=2.0.0
torch>=2.0
pytorch-lightning>=2.0
omegaconf
librosa
cython
